% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_finalWeights}
\alias{est_finalWeights}
\title{Estimate the Weights that Maximize the Probability of Observed Outcomes}
\usage{
est_finalWeights(y, ftk, sd, weights, threshold)
}
\arguments{
\item{y}{A numeric vector that indicates observed election outcomes.}

\item{ftk}{A numeric vector that indicates the mean of the forecast models. The models are assumed to follow a normal distribution.}

\item{sd}{A value that indicates the standard deviation of the forecast models. In this function, it is assumed to be 1 and remains consistent across the models.}

\item{weights}{A numeric vector that indicates the initial weights assigned to the forecast models.}

\item{threshold}{A value that indicates the level of tolerance in the difference between the estimated probability at iteration j-1 the estimated probability at iteration j.}
}
\value{
A list with three elements:
 \item{final_weights}{The final weights that maximize the probability that an outcome t is best predicted by a series of forcast models, given a pre-defined tolerance threshold.}
 \item{final_density}{The final maximized probability.}
 \item{iterations}{The number of iterations the function goes through to reach the maximized probability.}
}
\description{
With initial parameters, this function iteratively calculates the weights that seek to maximize the probability that an outcome t is best predicted by a series of forecast models. The function stops at iteration j and return the final weights when the estimated probability at iteration j converges based on a pre-defined threshold of your choosing.
}
\note{
The value in Equation (1) is what this function seeks to maximize.

The function uses the \code{est_ztk()} function (see Equation 2) and the \code{update_weights()} function (see Equation 3) to iteratively update the weights.
}
\examples{
y <- c(0.1,0.2,0.5)
ftk <- seq(0,1,0.1)
sd <- 1
weights <- rep(1/length(ftk), length(ftk)) ## Intially assign equal weight to each model
est_finalWeights(y=y, ftk=ftk, weights=weights, sd=1, threshold=0.0001)

}
\seealso{
\code{get_num_den}, \code{update_weights}, \code{est_finalweights}
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
