% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_ztk}
\alias{est_ztk}
\title{Calculate the Latent Quantity ztk}
\usage{
est_ztk(y, ftk, sd, weights)
}
\arguments{
\item{y}{A numeric vector that indicates observed election outcomes.}

\item{ftk}{A numeric vector that indicates the mean of the forecast models. The models are assumed to follow a normal distribution.}

\item{sd}{A value that indicates the standard deviation of the forecast models. In this function, it is assumed to be 1 and remains consistent across the models.}

\item{weights}{A numeric vector that indicates the initial weights assigned to the forecast models.}
}
\value{
A matrix, with the entry (t, k) representing the probability that observation t was best predicted by model k.
}
\description{
Calculate the Latent Quantity ztk, which is the fraction between the weighted density for each observed outcome t and the sum of weighted density across k forecast models for each t.
}
\note{
Refer to the equation in Step 2.
}
\examples{
y <- c(0.1,0.2,0.5)
ftk <- seq(0,1,0.1)
sd <- 1
weights <- rep(1/length(ftk), length(ftk)) ## Intially assign equal weight to each model
est_ztk(y=y, ftk=ftk, weights=weights, sd=1)

}
\seealso{
\code{get_num_den}, \code{update_weights}, \code{est_finalweights}
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
