% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_num_den}
\alias{get_num_den}
\title{Calculate the Weighted Density for Observed Outcomes and the Sum of Weighed Density for Each Model}
\usage{
get_num_den(y, ftk, sd, weights)
}
\arguments{
\item{y}{A numeric vector that indicates observed election outcomes.}

\item{ftk}{A numeric vector that indicates the mean of the forecast models. The models are assumed to follow a normal distribution.}

\item{sd}{A value that indicates the standard deviation of the forecast models. In this function, it is assumed to be 1 and remains consistent across the models.}

\item{weights}{A numeric vector that indicates the initial weights assigned to the forecast models.}
}
\value{
A list with two elements:
 \item{numerator_matrix}{Each entry (t, k) documents the weighted probability of each election outcome t in each forcast model k.}
 \item{denominator_vector}{The sum of weighted density across k forecast models for each t.}
}
\description{
Calculate the weighted density for each observed outcome t and the sum of weighted density across k forecast models for each t.
}
\note{
Refer to the equation in Equation (2). The numerator matrix refers to the values in the numerator and the denominator vectore refers to the values in the deminator.

This is an intermediate function, which is mainly used to calculate the ztk matrix in Equation (2). See how to calculate the ztk matrix by referring to the function \code{est_zsk()}.
}
\examples{
y <- c(0.1,0.2,0.5)
ftk <- seq(0,1,0.1)
sd <- 1
weights <- rep(1/length(ftk), length(ftk)) ## Intially assign equal weight to each model
get_num_den(y=y, ftk=ftk, weights=weights, sd=1)

}
\seealso{
\code{est_ztk}, \code{update_weights}, \code{est_finalweights}
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
